+++ components/src/lib.rs
@@ -3,6 +3,7 @@
 
 mod accordion;
 mod button;
+mod canvas;
 mod drag_drop;
 mod dropdown;
 mod external_link;
@@ -19,6 +20,7 @@ mod tooltip;
 
 pub use accordion::*;
 pub use button::*;
+pub use canvas::*;
 pub use drag_drop::*;
 pub use dropdown::*;
 pub use external_link::*;
