+++ hooks/Cargo.toml
@@ -31,6 +31,7 @@ nokhwa = {version = "0.10.3", features = ["input-native"], optional = true }
 bytes = "1.3.0"
 winit = "0.28.2"
 euclid = "0.22.7"
+skia-safe = { version = "0.60.0", features = ["gl", "textlayout", "svg"] }
 
 [dev-dependencies]
 dioxus = { git = "https://github.com/DioxusLabs/dioxus", rev="49c5a5043a16fc82210af146c345793dd448e519", features = ["macro", "hooks", "hot-reload"]}
