+++ hooks/src/lib.rs
@@ -5,6 +5,7 @@ mod rope_editor;
 mod text_editor;
 
 mod use_animation;
+mod use_canvas;
 mod use_editable;
 mod use_focus;
 mod use_node;
@@ -17,6 +18,7 @@ pub use rope_editor::*;
 pub use text_editor::*;
 
 pub use use_animation::*;
+pub use use_canvas::*;
 pub use use_editable::*;
 pub use use_focus::*;
 pub use use_node::*;
